Move Ingest to analysis and rename IngestMarkdown
